requirements.txt
fastapi
uvicorn
openai
pytest
ruff
httpx
python-dotenv
